+++ lldb/test/Shell/Commands/command-disassemble-mixed.c
@@ -1,5 +1,7 @@
 // invalid mixed disassembly line
 
+// XFAIL: system-windows
+
 // RUN: %clang_host -g %s -o %t
 // RUN: %lldb %t -o "dis -m -n main" -o "exit" | FileCheck %s